[REF] html_editor: move selection plugin tests from utils tests
Tests concerning methods of the selection plugin were wrongly put in the
selection utils tests. This moves them to the selection plugin tests.

task-4809282
